--- packages/maskbook/src/plugins/RedPacket/Dashboard/index.tsx
@@ -1,9 +1,0 @@
-import type { Plugin } from '@dimensiondev/mask-plugin-infra'
-import { base } from '../base'
-
-const dashboard: Plugin.Dashboard.Definition = {
-    ...base,
-    init(signal) {},
-}
-
-export default dashboard
