+++ packages/maskbook/src/plugins/RedPacket/SNSAdaptor/RedPacketDialog.tsx
