+++ packages/maskbook/src/plugins/RedPacket/SNSAdaptor/RedPacketThemeOverride.ts
@@ -10,7 +10,7 @@ export const PluginRedPacketTheme = unstable_createMuiStrictModeTheme({
                     position: 'relative',
                     '&::after': {
                         position: 'absolute',
-                        backgroundImage: `url(${new URL('./UI/present-default.png', import.meta.url)})`,
+                        backgroundImage: `url(${new URL('./present-default.png', import.meta.url)})`,
                         top: 250,
                         width: 60,
                         height: 60,
