+++ packages/maskbook/src/plugins/RedPacket/SNSAdaptor/hooks/useRedPacketHistory.ts
@@ -1,5 +1,5 @@
 import { useAsyncRetry } from 'react-use'
-import { RedPacketRPC } from '../messages'
+import { RedPacketRPC } from '../../messages'
 import type { ChainId } from '@dimensiondev/web3-shared'
 
 export function useRedPacketHistory(address: string, chainId: ChainId) {
