+++ packages/maskbook/src/plugins/RedPacket/SNSAdaptor/present-dai.png
