+++ packages/maskbook/src/plugins/RedPacket/Worker/database.ts
@@ -1,18 +1,9 @@
-import type { RedPacketRecord, RedPacketRecordInDatabase, RedPacketHistory } from './types'
-import { RedPacketPluginID } from './constants'
-import { createPluginDatabase } from '../../database/Plugin/wrap-plugin-database'
+import type { RedPacketRecord, RedPacketRecordInDatabase, RedPacketHistory } from '../types'
+import { RedPacketPluginID } from '../constants'
+import { createPluginDatabase } from '../../../database/Plugin/wrap-plugin-database'
 import { omit } from 'lodash-es'
 
 export const RedPacketDatabase = createPluginDatabase<RedPacketRecordInDatabase>(RedPacketPluginID)
-
-export async function getRedPackets(ids: string[]) {
-    const records: RedPacketRecord[] = []
-    for (const id of ids) {
-        const record = await getRedPacket(id)
-        if (record) records.push(record)
-    }
-    return records
-}
 
 export async function getRedPacket(id: string) {
     const record = await RedPacketDatabase.get('red-packet', id)
