+++ packages/maskbook/src/plugins/RedPacket/Worker/index.ts
@@ -1,5 +1,6 @@
 import type { Plugin } from '@dimensiondev/mask-plugin-infra'
 import { base } from '../base'
+import '../messages'
 
 const worker: Plugin.Worker.Definition = {
     ...base,
