+++ packages/maskbook/src/plugins/RedPacket/Worker/services.ts
@@ -1,5 +1,5 @@
-import type { RedPacketRecord } from './types'
-import { RedPacketMessage } from './messages'
+import type { RedPacketRecord } from '../types'
+import { RedPacketMessage } from '../messages'
 import * as database from './database'
 import type { ChainId } from '@dimensiondev/web3-shared'
 import * as subgraph from './apis'
@@ -15,9 +15,9 @@ export async function discoverRedPacket(record: RedPacketRecord) {
 }
 
 export async function getRedPacketHistoryWithPassword(address: string, chainId: ChainId) {
-    const historys = await subgraph.getRedPacketHistory(address, chainId)
-    const historysWithPassword = []
-    for (const history of historys) {
+    const histories = await subgraph.getRedPacketHistory(address, chainId)
+    const historiesWithPassword = []
+    for (const history of histories) {
         await database.updateV1ToV2(history)
         const record = await database.getRedPacket(history.txid)
         if (history.chain_id === chainId && record) {
@@ -25,7 +25,7 @@ export async function getRedPacketHistoryWithPassword(address: string, chainId: 
         } else {
             history.payload.password = history.password = ''
         }
-        historysWithPassword.push(history)
+        historiesWithPassword.push(history)
     }
-    return historysWithPassword
+    return historiesWithPassword
 }
