+++ packages/maskbook/src/plugins/RedPacket/index.ts
@@ -9,12 +9,6 @@ registerPlugin({
             import.meta.webpackHot &&
             import.meta.webpackHot.accept('./SNSAdaptor/index', () => hot(import('./SNSAdaptor'))),
     },
-    Dashboard: {
-        load: () => import('./Dashboard'),
-        hotModuleReload: (hot) =>
-            import.meta.webpackHot &&
-            import.meta.webpackHot.accept('./Dashboard/index', () => hot(import('./Dashboard'))),
-    },
     Worker: {
         load: () => import('./Worker'),
         hotModuleReload: (hot) =>
