+++ packages/maskbook/src/plugins/RedPacket/messages.ts
@@ -4,8 +4,12 @@ import { createPluginRPC } from '../utils/createPluginRPC'
 
 export interface RedPacketMessages {
     redPacketUpdated: void
-    rpc: unknown
+    _: unknown
 }
 if (import.meta.webpackHot) import.meta.webpackHot.accept()
 export const RedPacketMessage = createPluginMessage<RedPacketMessages>(RedPacketPluginID)
-export const RedPacketRPC = createPluginRPC(RedPacketPluginID, () => import('./services'), RedPacketMessage.events.rpc)
+export const RedPacketRPC = createPluginRPC(
+    RedPacketPluginID,
+    () => import('./Worker/services'),
+    RedPacketMessage.events._,
+)
