--- packages/maskbook/src/plugins/RedPacket/pipes.ts
@@ -1,9 +1,0 @@
-import { RedPacketStatus } from './types'
-
-export function resolveRedPacketStatus(listOfStatus: RedPacketStatus[]) {
-    if (listOfStatus.includes(RedPacketStatus.claimed)) return 'Claimed'
-    if (listOfStatus.includes(RedPacketStatus.refunded)) return 'Refunded'
-    if (listOfStatus.includes(RedPacketStatus.expired)) return 'Expired'
-    if (listOfStatus.includes(RedPacketStatus.empty)) return 'Empty'
-    return ''
-}
